requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mysql-connector==2.2.9
python-dotenv==0.19.0
websockets==9.1
yarl==1.6.3
Pillow~=8.3.1
mysql-connector-python~=8.0.26
requests~=2.26.0